% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcanvas-package.R
\name{grapes-plus-grapes}
\alias{grapes-plus-grapes}
\alias{\%+\%}
\title{String concatenation operator}
\usage{
a \%+\% b
}
\arguments{
\item{a, b}{Character vectors to concatenate.}
}
\value{
A character vector of \code{paste0(a, b)}.
}
\description{
Convenience infix wrapper around \code{\link[=paste0]{paste0()}}.
}
\keyword{internal}
