% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.screeplot}
\alias{pca.screeplot}
\title{Scree plot of PCA eigenvalues or PC variance}
\usage{
pca.screeplot(pc = NULL, eval = NULL, mode = "eval", pc.use = "all")
}
\arguments{
\item{pc}{A PCA result list (output of \code{\link[=pca]{pca()}}).}

\item{eval}{Optional eigenvalue vector if \code{pc} is not provided.}

\item{mode}{Either \code{"eval"} (eigenvalues) or \code{"pc"} (proportion variance).}

\item{pc.use}{Subset of PCs to plot: \code{"all"} or an integer vector.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots eigenvalues (or proportion of variance) against PC index.
}
