% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{quantvec}
\alias{quantvec}
\title{Quantile bucket labels}
\usage{
quantvec(x, n = 4L)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Integer, number of quantile bins (default 4 = quartiles).}
}
\value{
A factor of length \code{length(x)}.
}
\description{
Assigns each value of \code{x} to one of \code{n} equal-probability bins, returning a
factor with levels \verb{"1st", "2nd", "3rd", "4th", ...}.
}
